+++ com/pdfjet/RadioButton.java
@@ -127,7 +127,7 @@ public class RadioButton implements Drawable {
         page.drawCircle(x + r1 + penWidth, yBox + r1 + penWidth, r1);
 
         if (this.selected) {
-            page.drawCircle(x + r1 + penWidth, yBox + r1 + penWidth, r2, PathOperator.FILL);
+            page.fillCircle(x + r1 + penWidth, yBox + r1 + penWidth, r2);
         }
 
         // A linked label is blue.